New implementation for sample plans (#243)
* add helper functions for detection algorithm

* style: [CI] format

* algorithms necessary functions finish, integration testing not completely done

* delete trailing lines

* add test

* finish all implementation of sample plan, unit tested through notebook

* end to end flow succeed, integration testing ongoing

* thorough run of videos

* fix ego exit segment

* fix small bug

* fix detection problem after ego exit fix

* style: [CI] format

* add doc string

* revert data files

* .

* typing

* annotate typing

* add necessary typing

* style: [CI] format

* add ego stop action

* fix import

* Integrate detection estimation to pipeline (#186)

* .

* style: [CI] format

* split yolo and strongsort

* add detection estimation

* style: [CI] format

* style: [CI] format

* run detection estimation with pipeline

* style: [CI] format


* fix errors

* fix error

* clean up

* style: [CI] format

* detection estimation as stage

* style: [CI] format

* add detection_estimation to notebook

* style: [CI] format

* use logger instead of print

* style: [CI] format

* fix typing in python3.8

* style: [CI] format

* clean up

* style: [CI] format

* add TODO for detection coversio

* Add 3d calibration against the road segment (#187)

* add to do of car_loc3d conversion

* tested, including segment line in the road segment

* add location calibration

* style: [CI] format

* fix flake errors

* Update __init__.py


* benchmark

* fix LineString error + output trajectories for benchmark

* style: [CI] format

* performance benchmark

* Detection Estimation: add 3D detection and output DetectionInfo (#189)

* add 3d detection and DetectionInfo

* style: [CI] format

* fix numpy operations

* update nb

* style: [CI] format

* document 3d detection dimensions

* style: [CI] format

* fix lint error


* fix algorithms with new discovery in segment lines (#188)

* fix algorithms with new discovery in segment lines

* style: [CI] format

* include intersection in detection info so that it could be passed down

* style: [CI] format

* fix lint error


* cleanup duplicate road segments

* style: [CI] format

* allow none uid

* fix test

* fix linting error

* video with skipped frames

* style: [CI] format

* Construct segment trajectory (#191)

* add to do of car_loc3d conversion

* tested, including segment line in the road segment

* add location calibration

* style: [CI] format

* fix bug in intersection extended lines

* add stage of constructing segment trajectory

* style: [CI] format

* add docstring

* fix docstring

* style: [CI] format

* fix function naming

* style: [CI] format

* fix import

* style: [CI] format

* typing

* style: [CI] format

* add complete story algorithm

* style: [CI] format

* initial push of the segment trajectory stage

* style: [CI] format

* integrate segment trajectory to pipeline (#192)

* integrate to pipeline

* style: [CI] format

* fix type

* run benchmark

* style: [CI] format

* clean up and update instructions

* clean up playground

* style: [CI] cleanup python notebooks

* update dir

* style: [CI] cleanup python notebooks

* fix errors

* style: [CI] cleanup python notebooks

* style: [CI] format


* fix lint error


* add detection type annotation

* fix detection estimation errors (#195)

* fix detection estimation errors

* style: [CI] cleanup python notebooks

* style: [CI] format


* Add object type filter (#196)

* add type filter

* style: [CI] format


* Fix Detection Estimation segment mapping (#197)

* fix polygon contains

* style: [CI] format

* fix postgis LineString[] typing

* style: [CI] cleanup python notebooks


* pruning testing of detection estimation (#199)

* pruning testing of detection estimation

* style: [CI] format

* add type

* style: [CI] format


* fix the pipeline notebook

* benchmark on front cameras

* Optimized mapping (#200)

* new mapping

* style: [CI] format

* disable pruning and fix pipeline

* fix detection estimation pipeline for new mapping

* style: [CI] cleanup python notebooks

* new mapping benchmark

* clean up comments

* style: [CI] cleanup python notebooks

* style: [CI] format

* enable cache in strongsort

* style: [CI] format

* add customized predicate for detection_estimation to prune detection, and add fig14 pipeline

* style: [CI] cleanup python notebooks

* style: [CI] format

---------


* Fix road type (#201)

* fix the pipeline notebook

* benchmark on front cameras

* fix road type

* style: [CI] cleanup python notebooks

---------


* finish final merge

* commit investigation

* cmt investigation

* working investigation

* style: [CI] cleanup python notebooks

* style: [CI] format

* finish more debugging

* finish more investigation

* style: [CI] format

* style: [CI] cleanup python notebooks

* style: [CI] generate video examples

* style: [CI] generate video examples

* style: [CI] generate video examples

* commit latest test 0655

* style: [CI] cleanup python notebooks

* style: [CI] generate video examples

* style: [CI] generate video examples

* first commit to new approach

* style: [CI] format

* style: [CI] generate video examples

* style: [CI] generate video examples

* style: [CI] generate video examples

* proved working on 0655

* style: [CI] cleanup python notebooks

* style: [CI] format

* style: [CI] generate video examples

* clean up pr

* more clean up of the pr

* add back video

* fix difference

* fix fig13 end to end notebook

* no need to get ego road polygon

* style: [CI] format

* fix fps to 12

* fix predicate

* fix type

* lint

---------

+++ optimized_ingestion/camera_config.py
@@ -65,8 +65,8 @@ class CameraConfig:
     _data: "npt.NDArray[np.float32]"
 
     @property
-    def frame_num(self) -> float:
-        return self._data[0].item()
+    def frame_num(self) -> int:
+        return int(self._data[0].item())
 
     @property
     def camera_translation(self) -> Float3:
